pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "octodns-gitops"
version = "0.1.0"
description = "GitOps tooling for OctoDNS: processors, CLI wrappers, and workflow automation"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Serge van Ginderachter", email = "serge@ginsys.eu" },
]
keywords = ["dns", "octodns", "gitops", "infrastructure"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "octodns>=1.0",
    "octodns-hetzner",
    "pyyaml>=6.0",
    "tabulate>=0.9",
    "dnspython>=2.0",
]

[project.urls]
Homepage = "https://github.com/ginsys/octodns-gitops"
Repository = "https://github.com/ginsys/octodns-gitops"
Issues = "https://github.com/ginsys/octodns-gitops/issues"

[project.scripts]
octodns-gitops-validate = "octodns_gitops.cli.validate:main"
octodns-gitops-sync = "octodns_gitops.cli.sync:main"
octodns-gitops-drift = "octodns_gitops.cli.drift:main"
octodns-gitops-report = "octodns_gitops.cli.report:main"
octodns-gitops-init = "octodns_gitops.bootstrap.makefile:main"

[tool.hatch.build.targets.wheel]
packages = ["src/octodns_gitops"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/templates",
    "/README.md",
    "/LICENSE",
    "/NOTICE",
]
